requirements.txt
Django==3.0.7
Pillow==7.1.2
django-cleanup==5.0.0
PyMySQL==0.9.3
djangorestframework==3.11.0
django-cors-headers==3.4.0